// SDKs/Widget SDKs/Infor_WidgetSDK_3.0.1/samples/infor.sample.workspace/manifest-types.d.ts
// NOTE: This file has been automatically generated. It should never be manually edited.
import { ILanguage } from "@infor-lime/core";
export interface IManifestLanguage extends ILanguage<IManifestLanguage> {
	/** Edit User */
	editUser: string;
	/** Email */
	email: string;
	/** First Name */
	firstName: string;
	/** Last Name */
	lastName: string;
	/** This is where you would launch the application */
	launchToastMessage: string;
	/** User Workspace Launch */
	launchToastTitle: string;
	/** The details of the user were changed. */
	submitToastMessage: string;
	/** User Edited */
	submitToastTitle: string;
	/** Title */
	title: string;
	/** User Details */
	userDetails: string;
	/** Workspace sample widget */
	widgetDescription: string;
	/** Workspace Sample */
	widgetTitle: string;
}
export type IManifestSettings = Record<string, never>;